==> part_4.ipynb <==
import pandas as pd ## Обновить pandas до последней версии
import numpy as np
import math
import os
import glob
import mysql.connector
import datetime
from mysql.connector import errorcode
from sklearn import preprocessing
from sklearn import metrics

import matplotlib.pyplot as plt
import pandas.core.algorithms as algos
from pandas import Series
import scipy.stats.stats as stats
import re
import traceback
from datetime import date
import string

from os import listdir
from os.path import isfile, join


import warnings
warnings.filterwarnings("ignore")

user = 'project_user'
database = 'project_db'

my_path = 'C:\\Users\\User\\Documents\\Python Scripts\\IT\\Untitled Folder' 

#### Блок Калинина Артёма

scores_for_vars = pd.read_csv(my_path + '/output_3/scores_for_vars.csv')
scores_for_vars = scores_for_vars.drop(['Unnamed: 0'], axis = 1)

scores_for_vars.head(5)

woe_iv_table = pd.read_csv(my_path + '/output_3/woe_iv_table.csv')
woe_iv_table = woe_iv_table.drop(['Unnamed: 0'], axis = 1)

## Веса определяются по IV

IV_table = woe_iv_table[['VAR_NAME', 'IV']].drop_duplicates().copy()
IV_table = IV_table.reset_index()
IV_table = IV_table.drop(['index'], axis = 1)
IV_table

IV_table['predicive_power'] = np.where((IV_table.IV < 0.02), 'useless',
                                       np.where((IV_table.IV >= 0.02) & (IV_table.IV <= 0.1), 'weak',
                                               np.where((IV_table.IV > 0.1) & (IV_table.IV <= 0.3), 'medium',
                                                       np.where((IV_table.IV > 0.3) & (IV_table.IV <= 0.5), 'strong', 'suspicious'))))
                                       

IV_table

##### Предоставим пользовтелю задать вес для переменной с strong predictive power, остальные веса подсчитаются автоматически, при этом medium и weak являются релевантными, useless слабый вес - остальное на suspicious

IV_table.predicive_power.value_counts()

strong = float(input())##Минимильный вес равен 0.12

n_medium = 4
n_weak = 3
n_useless = 3
n_strong = 2
n_suspicious = 1

medium = (1 - strong * n_strong)/1.5/n_medium
weak = (1 - strong * n_strong - n_medium * medium)/n_weak
useless = (1 - strong * n_strong - n_medium * medium  - n_weak * weak)/n_useless
suspicious = (1 - strong * n_strong - n_medium *  medium - n_weak * weak - n_useless * useless)/n_suspicious
if medium > strong:
    print("Выберите вес для strong выше")

print(' strong=', strong, '\n', 'medium=',  "{0:.2f}".format(medium), '\n', 'weak=', "{0:.2f}".format(weak), '\n',  
      'useless=', "{0:.2f}".format(useless), '\n',  'suspicious=', "{0:.2f}".format(suspicious))
#Такие веса у каждой группы, то есть у каждой группы medium вес 0.11, поэтому данном принте сумма не будет 1

print(strong * n_strong + weak * n_weak + n_medium * medium + useless * n_useless + n_suspicious * suspicious)

IV_table

scores_for_vars['Total_score'] = (scores_for_vars.age * strong + scores_for_vars.age_of_car * medium + 
                                  scores_for_vars.children * weak + scores_for_vars.education * weak + 
                                  scores_for_vars.employed_by * suspicious + scores_for_vars.family * medium + 
                                  scores_for_vars.gender * useless + scores_for_vars.house_ownership * useless + 
                                  scores_for_vars.housing * useless + scores_for_vars.income * medium + 
                                  scores_for_vars.income_type * weak + scores_for_vars.marital_status * strong + 
                                  scores_for_vars.position * medium)

scores_for_vars['Total_score'].describe()

scores_for_vars.head()
